+++ libs/core/src/SwitchButton/SwitchButton.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 
 import { StyleSheet, Switch, Text, View } from 'react-native';
 
-import Icon, { icons } from './Icon';
+import Icon, { icons } from '../IconComp/Icon';
 
 type SwitchButtonProps = {
   icon: keyof typeof icons;
