+++ libs/core/src/icons/aboutMe.png
