+++ libs/core/src/icons/add.png
