+++ libs/core/src/icons/appreciation.png
