+++ libs/core/src/icons/arrow.png
