+++ libs/core/src/icons/attachment.png
