+++ libs/core/src/icons/blueAdd.png
