+++ libs/core/src/icons/call.png
