+++ libs/core/src/icons/camera.png
