+++ libs/core/src/icons/chat.png
