+++ libs/core/src/icons/comments.png
