+++ libs/core/src/icons/connection.png
