+++ libs/core/src/icons/darkMode.png
