+++ libs/core/src/icons/delete.png
