+++ libs/core/src/icons/design.png
