+++ libs/core/src/icons/done.png
