+++ libs/core/src/icons/dot.png
