+++ libs/core/src/icons/edit.png
