+++ libs/core/src/icons/education.png
