+++ libs/core/src/icons/eyeIcon.png
