+++ libs/core/src/icons/filter.png
