+++ libs/core/src/icons/finance.png
