+++ libs/core/src/icons/google.png
