+++ libs/core/src/icons/headhunting.png
