+++ libs/core/src/icons/health.png
