+++ libs/core/src/icons/home.png
