+++ libs/core/src/icons/language.png
