+++ libs/core/src/icons/languageAddIcon.png
