+++ libs/core/src/icons/location.png
