+++ libs/core/src/icons/logout.png
