+++ libs/core/src/icons/notifications.png
