+++ libs/core/src/icons/offNotifications.png
