+++ libs/core/src/icons/options.png
