+++ libs/core/src/icons/orangeSearch.png
