+++ libs/core/src/icons/password.png
