+++ libs/core/src/icons/pdf.png
