+++ libs/core/src/icons/picture.png
