+++ libs/core/src/icons/programmer.png
