+++ libs/core/src/icons/redDelete.png
