+++ libs/core/src/icons/redHeart.png
