+++ libs/core/src/icons/remove.png
