+++ libs/core/src/icons/restaurant.png
