+++ libs/core/src/icons/resume.png
