+++ libs/core/src/icons/save.png
