+++ libs/core/src/icons/search.png
