+++ libs/core/src/icons/send.png
