+++ libs/core/src/icons/settings.png
