+++ libs/core/src/icons/share.png
