+++ libs/core/src/icons/skill.png
