+++ libs/core/src/icons/upArrow.png
