+++ libs/core/src/icons/uploadCv.png
