+++ libs/core/src/icons/work.png
