+++ libs/core/src/icons/write.png
